requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
celery==5.3.0
redis==4.5.2
redlock-py